Refactor model selection and context management for improved functionality
- Replaced the `Select` component in `ModelComparisonCard` with `ModelSelectorBase` for enhanced model selection experience, including filtering capabilities.
- Introduced `ModelsContext` to manage search queries, sorting, and filters across the models page, improving state management and reusability.
- Updated `HomePage` to utilize `ModelsProvider`, streamlining the handling of model data and user interactions.
- Removed redundant state management from `HomePage`, leveraging the new context for cleaner code.

These changes enhance the user experience by providing a more intuitive model selection process and better state management across components.
